changed constant to double

+++ tp3/src/main/java/Main.java
@@ -2,7 +2,7 @@ import java.io.File;
 import java.util.ArrayList;
 
 public class Main {
-    private static final int[][] logicNumbers = {{1,1},{1,-1},{-1,1},{-1,-1}};
+    private static final double[][] logicNumbers = {{1,1},{1,-1},{-1,1},{-1,-1}};
 
     public static void main(String[] args) {
         File file = new File("TP3-ej2-Conjunto-entrenamiento.txt");
